% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horsekick.R
\name{clean_horse_kicks}
\alias{clean_horse_kicks}
\title{Removes and Warns if Horse Kick Data Malformed}
\usage{
clean_horse_kicks(horse_kick_df)
}
\arguments{
\item{horse_kick_df}{A dataframe with at least two columns: year and corps}
}
\value{
A dataframe with two columns: year and corps and only valid rows
}
\description{
The model needs a year range between 1701-1919.
The model is only aware of I-XV and G corps.
}
