% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horsekick.R
\name{predict_horse_kicks}
\alias{predict_horse_kicks}
\title{Predict the Number of Horse Kicks in a Prussian Corps}
\usage{
predict_horse_kicks(clean_kick_df)
}
\arguments{
\item{clean_kick_df}{A data frame generated from \code{clean_horse_kicks}}
}
\value{
A column of of predictions
}
\description{
Predict the Number of Horse Kicks in a Prussian Corps
}
